+++ packages/core/src/source/index.ts
@@ -1,7 +1,6 @@
 export * from './source';
 export * from './loader';
 export { getSlugs } from './plugins/slugs';
-export { statusBadgesPlugin, type StatusBadgeItem } from './plugins/status-badges';
 export { FileSystem } from './storage/file-system';
 export * as PathUtils from './path';
